Add TO Go GET federations
